IS-1975: Refactor behandlerdialog folder into /sider folder
