+++ src/sider/behandlerdialog/meldinger/MeldingInnholdPanel.tsx
@@ -7,14 +7,14 @@ import { BodyLong, Detail, Panel } from "@navikt/ds-react";
 import { PaperclipIcon } from "@navikt/aksel-icons";
 import styled from "styled-components";
 import { tilDatoMedManedNavnOgKlokkeslett } from "@/utils/datoUtils";
-import { VisMelding } from "@/components/behandlerdialog/meldinger/VisMelding";
-import PdfVedleggLink from "@/components/behandlerdialog/meldinger/PdfVedleggLink";
+import { VisMelding } from "@/sider/behandlerdialog/meldinger/VisMelding";
+import PdfVedleggLink from "@/sider/behandlerdialog/meldinger/PdfVedleggLink";
 import { DocumentComponentType } from "@/data/documentcomponent/documentComponentTypes";
-import { PaminnelseWarningIcon } from "@/components/behandlerdialog/paminnelse/PaminnelseWarningIcon";
+import { PaminnelseWarningIcon } from "@/sider/behandlerdialog/paminnelse/PaminnelseWarningIcon";
 import { getHeaderText } from "@/utils/documentComponentUtils";
 import { useVeilederInfoQuery } from "@/data/veilederinfo/veilederinfoQueryHooks";
 import { meldingTypeTexts } from "@/data/behandlerdialog/behandlerdialogTexts";
-import { ReturLegeerklaringWarningIcon } from "@/components/behandlerdialog/legeerklaring/ReturLegeerklaringWarningIcon";
+import { ReturLegeerklaringWarningIcon } from "@/sider/behandlerdialog/legeerklaring/ReturLegeerklaringWarningIcon";
 
 const MeldingTekstWrapper = styled(BodyLong)`
   white-space: pre-wrap;
