+++ src/sider/behandlerdialog/meldinger/SamtaleAccordionItem.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { MeldingDTO } from "@/data/behandlerdialog/behandlerdialogTypes";
 import { tilDatoMedManedNavnOgKlokkeslett } from "@/utils/datoUtils";
 import { Accordion } from "@navikt/ds-react";
-import { MeldingerISamtale } from "@/components/behandlerdialog/meldinger/MeldingerISamtale";
+import { MeldingerISamtale } from "@/sider/behandlerdialog/meldinger/MeldingerISamtale";
 import { StetoskopIkon } from "../../../../img/ImageComponents";
 import { useBehandlerNavn } from "@/hooks/behandler/useBehandlerNavn";
-import { SamtaleTags } from "@/components/behandlerdialog/meldinger/SamtaleTags";
+import { SamtaleTags } from "@/sider/behandlerdialog/meldinger/SamtaleTags";
 import { FlexRow } from "@/components/Layout";
 
 interface SamtaleAccordionItemProps {
