+++ src/sider/behandlerdialog/meldingtilbehandler/MeldingsTypeInfo.tsx
