+++ src/sider/behandlerdialog/paminnelse/PaminnelseWarningIcon.tsx
