+++ test/behandlerdialog/MeldingerTest.tsx
@@ -5,7 +5,7 @@ import { navEnhet } from "../dialogmote/testData";
 import React from "react";
 import { queryClientWithMockData } from "../testQueryClient";
 import { expect } from "chai";
-import { Meldinger } from "@/components/behandlerdialog/meldinger/Meldinger";
+import { Meldinger } from "@/sider/behandlerdialog/meldinger/Meldinger";
 import { behandlerdialogQueryKeys } from "@/data/behandlerdialog/behandlerdialogQueryHooks";
 import {
   ARBEIDSTAKER_DEFAULT,
